Arona Airfield (1.0.0) XP12
Hello, this is my newest release of my scenery development for the Lago Maggiore AIRTaxi Series 

This is now for the southern part of the Lago near the city of Arona, ITA. 

My favorite plane to to AIRTaxi operation is the Cirrus Vision SF50! But sometime it can be hard to land it so that you don't go over the runway end!

I highly recommend to download satellite imagery for this scenery it just looks a lot better!

Make sure to download the other scenery's via the Google drive!

And make sure to watch my Showcase series of how to operate at the Lago Maggiore: 

https://youtu.be/Byv4mx0iwFE

https://youtu.be/viCMutMPvkA

(libraries are also available on the Lago Maggiore AIRTaxi main page)

I Hope you enjoy it and have fun!