Intra Airfield (V1.0.2) XP12-11
Hi, after a quite some time here is the next part for the Lago Maggiore AIRTaxi series!

This time its Intra Airfield, Intra is located on the western side of the Lago Maggiore. (Directly across from Laveno.)

This Airfield has a 600m long runway and some other objects on the ramp. But stay tuned for the Terminal it will be completely modeled and textured by me and in my opinion it looks pretty good. It will come in the next few Updates.

Hope you like it and DONT forget to download the other Parts of the Lago Maggiore AIRTaxi series!

And the scenery is also with Xplane 11 compatable.

But for the scenerys to work you will NEED to download the libraries from the ,,libraries.txt''